% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priestley.R
\docType{data}
\name{Biographies}
\alias{Biographies}
\alias{Specimen}
\title{Priestley's Chart of Biography Data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2418 rows and 26 columns.}
\usage{
Biographies

Specimen
}
\description{
These are several datasets from the Joseph Priestley's *A Chart of Biography*,
which is the first known use of a visual timeline to display historical data.
The data consist of the lives of approximately 2,000 individuals between 1200 B.C.E.
and about 1800 C.E. The *Chart of Biography* was published in numerous
editions between 1764 and 1815. This package provides the data associated
with the 1st (1764) and 74th (1778) editions.x
}
\details{
These datasets comprise the biographical data in Priestley's
*Chart of Biography*.
}
\keyword{datasets}
