% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priestley.R
\docType{data}
\name{Kings}
\alias{Kings}
\title{List of Rulers, 1300 BCE to 1800}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 164 rows and 3 columns.}
\usage{
Kings
}
\description{
The data frame `Kings` is a list of rulers: Kings of Judah, King of Babylon,
which are used to annotated the margin of the chart. These data are from
the list titled "The Times When Kings in Those Successions Which Are Noted
in the Margin of the Chart Began Their Reigns".
These data are from the 1st Edition (1764), pp. 77--80.
}
\keyword{datasets}
