% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Specimen.R
\docType{data}
\name{Specimen}
\alias{Specimen}
\title{List of persons in Priestley's A Specimen of A Chart of Biography}
\format{A data frame with 59 rows and 8 columns.
\describe{
\item{row}{(integer) The row (from top to bottom, within each `division`) which the individual appears in the chart.}
\item{name}{(string) The name used for the individual in the "Specimen" chart. This may be different than the one used in the name index.}
\item{text}{(string) Original text from name index.}
\item{division}{(string) In the Specimen chart, individuals are divided into two broad classes of occupations: "Statesmen" and "Men of Learning".}
\item{born_min}{(integer) Lower estimate of the birth year of an individual. This is used as the start of the uncertain segment of a lifespan on the timeline. Negative numbers are years BCE (0 = 1 BCE, -1 = 2 BCE).}
\item{born_max}{(integer) Upper estimate of the birth year of an individual. This is used as the start of the certain segment of a lifespan on the timeline. Negative numbers are years BCE (0 = 1 BCE, -1 = 2 BCE).}
\item{died_min}{(integer) Lower estimate of the death year of an individual. This is used as the end of the certain segment of a lifespan on the timeline. Negative numbers are years BCE (0 = 1 BCE, -1 = 2 BCE).}
\item{died_max}{(integer) Upper estimate of the death year of an individual. This is used as the end of the uncertain segment of a lifespan on the timeline. Negative numbers are years BCE (0 = 1 BCE, -1 = 2 BCE).}
}}
\usage{
Specimen
}
\description{
The "Specimen" only contains ... individuals, divided into two
categories ("Statesmen" and "Men of Learning") rather than six.
Otherwise, the data is identical to that in `Biographies`.
}
\keyword{datasets}
